Two bugs.
Resolved issue with throwing an exception when editing customer birthday when supplied string is 0000-00-00 00:00:00 (it is a valid date in the DB and thus valid birthday).
No longer have to supply hardcoded event range for building customer and checkins table. It will look in the DB to get all events that are currently on and use those.